requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.4.0
